pyproject.toml
[project]
name = "dsmr2mqtt"
version = "4.0.2"
description = "MQTT client for Belgian and Dutch Smart Meter (DSMR) - Slimme Meter"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.12"
authors = [
    { name = "Hans IJntema" }
]
keywords = ["dsmr", "mqtt", "smart-meter", "home-assistant", "energy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Home Automation",
]
dependencies = [
    "paho-mqtt>=2.0.0",
    "pyserial>=3.5",
    "persist-queue>=0.8.0",
    "packaging>=23.0",
    "structlog>=25.0.0",
]

[project.urls]
Homepage = "https://github.com/thijsmie/dsmr2mqtt"
Repository = "https://github.com/thijsmie/dsmr2mqtt"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/dsmr2mqtt"]

[project.scripts]
dsmr2mqtt = "dsmr2mqtt.__main__:main"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["dsmr2mqtt"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
